+++ CloudShell.Abstractions/ResourceManager/ResourceStandardViewIds.cs
@@ -1,0 +1,14 @@
+namespace CloudShell.Abstractions.ResourceManager;
+
+public static class ResourceStandardViewIds
+{
+    public const string Overview = "overview";
+    public const string Configuration = "configuration";
+    public const string Endpoints = "endpoints";
+    public const string Dns = "dns";
+    public const string Identity = "identity";
+    public const string Volumes = "volumes";
+    public const string Activity = "activity";
+    public const string Environment = "environment";
+    public const string Storage = "storage";
+}